refactor: use central estimate service
